Don't draw silk on the far side of the board
Silk is within pcb_max_layer, which makes it look like a 'normal'
layer and get caught by the code which checks whether a normal layer
should be drawn. However, there are two silk layers, and we always want to
skip drawing silk for the far side. By moving the check for silk
before the normal layer code, we can have that special case run.

Signed-off-by: Keith Packard <keithp@keithp.com>


git-svn-id: svn://repo.hu/librnd/trunk@7527 1d61de99-560f-4dba-8957-1e505d42358b